src/main/java/com/codingshadows/homework/SerializationUtils.java
package com.codingshadows.homework;

import java.io.*;

public class SerializationUtils {
    public static void serializeObject(String filePath, Serializable object) throws IOException {
        try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(filePath))) {
            oos.writeObject(object);
        }
    }

    public static Object deserializeObject(String filePath) throws IOException, ClassNotFoundException {
        try (ObjectInputStream ois = new ObjectInputStream(new FileInputStream(filePath))) {
            return ois.readObject();
        }
    }
}
